Added "show variants" button.
Button will become active when a sample name is selected from the sample data table.  Once clicked the app will switch to view the variant tab for the selected sample.  Button will de-activate and variant tab will hide when the sample is un-selected and variant tab hides when a new sample is selected so user must hit the button again.