Added SelectedSamples().  Can select multiple samples on the sample data table.
Multiple samples can be selected correctly from the sample data table.  The eventReactive SelectedSamples() will return the matrix of indices of selected samples so that the "show variants" button correctly enables/disables.  Next step is to pass the selected sample names onto GetVCF() and PlotCoverage() and to render the correct number of VCF tables and coverage plots on-the-fly.  Likely need uiOutput() & renderUI().