Modifications to common variant functions.
Changed the "Variant" class to hold a data frame of information about each variant to be displayed in the common variant table.  Made changes to BuildSampleObjects() and FindExactMatches() to incorporate this additional information.  Still need to put in a way to handle the search for exact matches when all the selected samples have no data in their VCF files.